Fixed Bug, Started Enemy Attack, Used AI
I fixed the bug that accelerates when running into a wall. Now, it will no longer accelerate when running into a wall. I began what would soon be the enemy attack. The squirrel will charge into the player but it does not work as of this update. Also, in October, I was experimenting with AI to see how it tries to fix the code, but it didn't work as I intended so it can be ignored.