Made Weapon Switch
If we add more weapons, the method I am using will be too complex for efficiency. You are able to switch to remove the sword weapon by picking up the ball_weapon on the players left in damage_test and you are able to get it back by picking up the sword_weapon on the platform. I did not disable acceleration or damage with movement when the sword weapon is in use though although that is what is intended.